+++ Snake/Directions.java
@@ -1,4 +1,4 @@
-package snake;
+package Snake;
 
 enum Directions {
     UP,
