+++ Snake/Game.java
@@ -1,4 +1,4 @@
-package snake;
+package Snake;
 
 abstract class Game {
     private static TileEnum[][] gridArray;
@@ -10,15 +10,5 @@ abstract class Game {
             }
         }
         return Game.gridArray;
-    }
-}
-enum TileEnum {
-    SNAKE('@'), EMPTY(' '), APPLE('*');
-    private final char c;
-    TileEnum(char c) {
-        this.c = c;
-    }
-    public char getValue() {
-        return this.c;
     }
 }
