+++ Snake/GameAction.java
@@ -1,0 +1,10 @@
+package Snake;
+
+enum GameAction {
+    MOVE_UP,
+    MOVE_DOWN,
+    MOVE_LEFT,
+    MOVE_RIGHT,
+    GAME_START,
+    GAME_QUIT
+}
