+++ Snake/Main.java
@@ -1,4 +1,4 @@
-package snake;
+package Snake;
 
 import static java.lang.System.exit;
 import static java.lang.Thread.sleep;
